Cleaning up internal API for memory functions
- Also Split up device.cpp and memory.cpp
