+++ src/backend/opencl/memory.cpp
@@ -88,7 +88,7 @@ size_t MemoryManager::getMaxMemorySize(int id)
 }
 
 MemoryManager::MemoryManager() :
-    common::MemoryManager(getDeviceCount(), MAX_BUFFERS, AF_MEM_DEBUG || AF_OPENCL_MEM_DEBUG)
+    common::MemoryManager(getDeviceCount(), common::MAX_BUFFERS, AF_MEM_DEBUG || AF_OPENCL_MEM_DEBUG)
 {
     this->setMaxMemorySize();
 }
@@ -128,7 +128,7 @@ size_t MemoryManagerPinned::getMaxMemorySize(int id)
 }
 
 MemoryManagerPinned::MemoryManagerPinned() :
-    common::MemoryManager(getDeviceCount(), MAX_BUFFERS, AF_MEM_DEBUG || AF_OPENCL_MEM_DEBUG),
+    common::MemoryManager(getDeviceCount(), common::MAX_BUFFERS, AF_MEM_DEBUG || AF_OPENCL_MEM_DEBUG),
     pinned_maps(getDeviceCount())
 {
     this->setMaxMemorySize();
@@ -184,6 +184,11 @@ size_t getMaxBytes()
     return getMemoryManager().getMaxBytes();
 }
 
+unsigned getMaxBuffers()
+{
+    return getMemoryManager().getMaxBuffers();
+}
+
 void garbageCollect()
 {
     getMemoryManager().garbageCollect();
@@ -197,18 +202,27 @@ void printMemInfo(const char *msg, const int device)
 template<typename T>
 T* memAlloc(const size_t &elements)
 {
-    return (T *)getMemoryManager().alloc(elements * sizeof(T));
+    return (T *)getMemoryManager().alloc(elements * sizeof(T), false);
 }
 
-cl::Buffer *bufferAlloc(const size_t &bytes)
+void* memAllocUser(const size_t &bytes)
 {
-    return (cl::Buffer *)getMemoryManager().alloc(bytes);
+    return getMemoryManager().alloc(bytes, true);
 }
-
 template<typename T>
 void memFree(T *ptr)
 {
     return getMemoryManager().unlock((void *)ptr, false);
+}
+
+void memFreeUser(void *ptr)
+{
+    getMemoryManager().unlock((void *)ptr, true);
+}
+
+cl::Buffer *bufferAlloc(const size_t &bytes)
+{
+    return (cl::Buffer *)getMemoryManager().alloc(bytes, false);
 }
 
 void bufferFree(cl::Buffer *buf)
@@ -216,24 +230,15 @@ void bufferFree(cl::Buffer *buf)
     return getMemoryManager().unlock((void *)buf, false);
 }
 
-template<typename T>
-void memFreeLocked(T *ptr, bool user_unlock)
-{
-    getMemoryManager().unlock((void *)ptr, user_unlock);
-}
-
-template<typename T>
-void memLock(const T *ptr)
+void memLock(const void *ptr)
 {
     getMemoryManager().userLock((void *)ptr);
 }
 
-template<typename T>
-void memUnlock(const T *ptr)
+void memUnlock(const void *ptr)
 {
     getMemoryManager().userUnlock((void *)ptr);
 }
-
 
 void deviceMemoryInfo(size_t *alloc_bytes, size_t *alloc_buffers,
                       size_t *lock_bytes,  size_t *lock_buffers)
@@ -245,7 +250,7 @@ void deviceMemoryInfo(size_t *alloc_bytes, size_t *alloc_buffers,
 template<typename T>
 T* pinnedAlloc(const size_t &elements)
 {
-    return (T *)getMemoryManagerPinned().alloc(elements * sizeof(T));
+    return (T *)getMemoryManagerPinned().alloc(elements * sizeof(T), false);
 }
 
 template<typename T>
@@ -257,9 +262,6 @@ void pinnedFree(T* ptr)
 #define INSTANTIATE(T)                                      \
     template T* memAlloc(const size_t &elements);           \
     template void memFree(T* ptr);                          \
-    template void memFreeLocked(T* ptr, bool user_unlock);  \
-    template void memLock(const T* ptr);                    \
-    template void memUnlock(const T* ptr);                  \
     template T* pinnedAlloc(const size_t &elements);        \
     template void pinnedFree(T* ptr);                       \
